refactor: altera atributo score de Double para BigDecimal em AvaliationEntity

+++ src/main/java/ctw/screenscoreapi/Avaliations/application/dtos/create/CreateAvaliationToEntity.java
@@ -1,8 +1,10 @@
 package ctw.screenscoreapi.Avaliations.application.dtos.create;
 
+import java.math.BigDecimal;
+
 public record CreateAvaliationToEntity(
         Long movieId,
         Long userId,
-        Double score,
+        BigDecimal score,
         String comment
 ) {}
